+++ includes/features/class-wpcli.php
@@ -33,7 +33,7 @@ use Decalog\Plugin\Feature\DLogger;
 use Decalog\Plugin\Feature\SDK;
 
 /**
- * Manages DecaLog, view events logs and send messages to loggers.
+ * Manages DecaLog, view events logs and send events to loggers.
  *
  * @package Features
  * @author  Pierre Lannoy <https://pierre.lannoy.fr/>.
@@ -141,6 +141,7 @@ class Wpcli {
 	 * @since   2.0.0
 	 */
 	private function error( $code = 255, $stdout = false ) {
+		$msg = '[' . DECALOG_PRODUCT_NAME . '] ' . ucfirst( $this->exit_codes[ $code ] );
 		if ( \WP_CLI\Utils\isPiped() ) {
 			// phpcs:ignore
 			fwrite( STDOUT, '' );
@@ -148,11 +149,11 @@ class Wpcli {
 			exit( $code );
 		} elseif ( $stdout ) {
 			// phpcs:ignore
-			fwrite( STDERR, ucfirst( $this->exit_codes[ $code ] ) );
+			fwrite( STDERR, $msg );
 			// phpcs:ignore
 			exit( $code );
 		} else {
-			\WP_CLI::error( $this->exit_codes[ $code ] );
+			\WP_CLI::error( $msg );
 		}
 	}
 
@@ -165,6 +166,7 @@ class Wpcli {
 	 * @since   2.0.0
 	 */
 	private function warning( $msg, $result = '', $stdout = false ) {
+		$msg = '[' . DECALOG_PRODUCT_NAME . '] ' . ucfirst( $msg );
 		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
 			// phpcs:ignore
 			fwrite( STDOUT, $result );
@@ -182,6 +184,7 @@ class Wpcli {
 	 * @since   2.0.0
 	 */
 	private function success( $msg, $result = '', $stdout = false ) {
+		$msg = '[' . DECALOG_PRODUCT_NAME . '] ' . ucfirst( $msg );
 		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
 			// phpcs:ignore
 			fwrite( STDOUT, $result );
@@ -1167,7 +1170,7 @@ class Wpcli {
 		switch ( $action ) {
 			case 'list':
 				if ( 'full' === $detail ) {
-					$detail = [ 'id', 'class', 'name', 'product', 'version', 'available', 'enabled' ];
+					$detail = [ 'id', 'class', 'name', 'product', 'version', 'available', 'enabled', 'step' ];
 				} else {
 					$detail = [ 'id', 'name', 'available', 'enabled' ];
 				}
@@ -1240,7 +1243,7 @@ class Wpcli {
 	 * <enable|disable>
 	 * : The action to take.
 	 *
-	 * <early-loading|auto-logging|auto-start|auth-endpoint>
+	 * <early-loading|auto-logging|auto-start|auth-endpoint|slow-query-warn|query-trace>
 	 * : The setting to change.
 	 *
 	 * [--yes]
@@ -1281,6 +1284,14 @@ class Wpcli {
 						Option::network_set( 'metrics_authent', true );
 						$this->success( 'endpoints authentication is now activated.', '', $stdout );
 						break;
+					case 'slow-query-warn':
+						Option::network_set( 'slow_query_warn', true );
+						$this->success( 'slow-query warning is now activated.', '', $stdout );
+						break;
+					case 'trace-query':
+						Option::network_set( 'trace_query', true );
+						$this->success( 'query tracing is now activated.', '', $stdout );
+						break;
 					default:
 						$this->error( 7, $stdout );
 				}
@@ -1307,6 +1318,16 @@ class Wpcli {
 						Option::network_set( 'metrics_authent', false );
 						$this->success( 'endpoints authentication is now deactivated.', '', $stdout );
 						break;
+					case 'slow-query-warn':
+						\WP_CLI::confirm( 'Are you sure you want to deactivate slow-query warning?', $assoc_args );
+						Option::network_set( 'slow_query_warn', false );
+						$this->success( 'slow-query warning is now deactivated.', '', $stdout );
+						break;
+					case 'trace-query':
+						\WP_CLI::confirm( 'Are you sure you want to deactivate query tracing?', $assoc_args );
+						Option::network_set( 'trace_query', false );
+						$this->success( 'query tracing is now deactivated.', '', $stdout );
+						break;
 					default:
 						$this->error( 7, $stdout );
 				}
@@ -1317,7 +1338,7 @@ class Wpcli {
 	}
 
 	/**
-	 * Send a message to all running loggers.
+	 * Send an event to all running loggers.
 	 *
 	 * ## OPTIONS
 	 *
@@ -1353,9 +1374,9 @@ class Wpcli {
 		if ( ! in_array( $level, [ 'info', 'notice', 'warning', 'error', 'critical', 'alert' ], true ) ) {
 			$this->error( 10, $stdout );
 		}
-		$logger = Log::bootstrap( 'core', 'WP-CLI', defined( 'WP_CLI_VERSION' ) ? WP_CLI_VERSION : 'x' );
+		$logger = Log::bootstrap( 'plugin', DECALOG_PRODUCT_SHORTNAME, DECALOG_VERSION );
 		$logger->log( $level, $message, $code );
-		$this->success( 'message sent.', 'OK', $stdout );
+		$this->success( 'event triggered and sent.', 'OK', $stdout );
 	}
 
 	/**
@@ -1749,7 +1770,7 @@ class Wpcli {
 	 *
 	 */
 	public function tail( $args, $assoc_args ) {
-		if ( ! function_exists( 'shmop_open' ) || ! function_exists( 'shmop_read' ) || ! function_exists( 'shmop_write' ) || ! function_exists( 'shmop_delete' ) || ! function_exists( 'shmop_close' ) ) {
+		if ( ! function_exists( 'shmop_open' ) || ! function_exists( 'shmop_read' ) || ! function_exists( 'shmop_write' ) || ! function_exists( 'shmop_delete' ) ) {
 			$this->error( 11 );
 		}
 		if ( ! Autolog::is_enabled() ) {