+++ includes/system/class-sharedmemory.php
@@ -64,7 +64,7 @@ class SharedMemory {
 	 * @since    1.0.0
 	 */
 	public static function init() {
-		self::$available = ( function_exists( 'shmop_open' ) && function_exists( 'shmop_read' ) && function_exists( 'shmop_write' ) && function_exists( 'shmop_delete' ) && function_exists( 'shmop_close' ) );
+		self::$available = ( function_exists( 'shmop_open' ) && function_exists( 'shmop_read' ) && function_exists( 'shmop_write' ) && function_exists( 'shmop_delete' ) );
 	}
 
 	/**
@@ -107,9 +107,9 @@ class SharedMemory {
 	 * @since    2.0.0
 	 */
 	private function exists() {
-		$result = is_resource( $this->acquire() );
+		$result = decalog_is_shmop_resource( $this->acquire() );
 		if ( $result ) {
-			shmop_close( $this->shmid );
+			decalog_shmop_close( $this->shmid );
 		}
 		return $result;
 	}
@@ -130,24 +130,24 @@ class SharedMemory {
 			$cpt = 0;
 			while ( 20 > $cpt ) {
 				$this->shmid = $this->acquire( 'w', $this->perms, 0 );
-				if ( is_resource( $this->shmid ) ) {
+				if ( decalog_is_shmop_resource( $this->shmid ) ) {
 					break;
 				} else {
 					$cpt++;
 					usleep( 100 );
 				}
 			}
-			if ( is_resource( $this->shmid ) ) {
+			if ( decalog_is_shmop_resource( $this->shmid ) ) {
 				shmop_delete( $this->shmid );
-				shmop_close( $this->shmid );
+				decalog_shmop_close( $this->shmid );
 			} else {
 				return false;
 			}
 		}
 		$this->shmid = $this->acquire( 'c', $this->perms, $size );
-		if ( is_resource( $this->shmid ) ) {
+		if ( decalog_is_shmop_resource( $this->shmid ) ) {
 			$result = shmop_write( $this->shmid, $data, 0 );
-			shmop_close( $this->shmid );
+			decalog_shmop_close( $this->shmid );
 			return $result;
 		}
 		return false;
@@ -168,17 +168,17 @@ class SharedMemory {
 			$cpt = 0;
 			while ( 20 > $cpt ) {
 				$this->shmid = $this->acquire( 'w', $this->perms, 0 );
-				if ( is_resource( $this->shmid ) ) {
+				if ( decalog_is_shmop_resource( $this->shmid ) ) {
 					break;
 				} else {
 					$cpt++;
 					usleep( 100 );
 				}
 			}
-			if ( is_resource( $this->shmid ) ) {
+			if ( decalog_is_shmop_resource( $this->shmid ) ) {
 				$size = shmop_size( $this->shmid );
 				$data = shmop_read( $this->shmid, 0, $size );
-				shmop_close( $this->shmid );
+				decalog_shmop_close( $this->shmid );
 			} else {
 				return [];
 			}